requirements.txt
django
djangorestframework
django-rest-auth
django-allauth
django-cors-headers
